requirements_realtime.txt
# Additional requirements for real-time communication features

# WebSocket support
Flask-SocketIO==5.3.6
python-socketio==5.9.0
eventlet==0.33.3

# Redis for WebSocket message handling
redis==5.0.1

# Background task processing
celery==5.3.4
kombu==5.3.4

# Real-time notifications
pushbullet.py==0.11.0
requests==2.31.0

# Email notifications
Flask-Mail==0.9.1
sendgrid==6.10.0

# Caching for performance
Flask-Caching==2.1.0

# Database connection pooling
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# JSON Web Tokens for WebSocket auth
PyJWT==2.8.0

# Environment variables
python-dotenv==1.0.0

# Async support
asyncio-mqtt==0.13.0
